+++ howdy/src/pam/README.md
@@ -1,20 +1,35 @@
 # Howdy PAM module
 
+## Prepare
+
+This module depends on `INIReader` and `libevdev`.
+They can be installed with these packages:
+
+```
+Arch Linux - libinih libevdev
+Debian     - libinih-dev libevdev-dev
+Fedora     - inih-devel libevdev-devel
+OpenSUSE   - inih libevdev-devel
+```
+
+If your distribution doesn't provide `INIReader`,
+it will be automatically pulled from git at the subproject's pinned version.
+
 ## Build
 
-```sh
-meson setup build -Dinih:with_INIReader=true
+``` sh
+meson setup build
 meson compile -C build
 ```
 
 ## Install
 
-```sh
-sudo mv build/libpam_howdy.so /lib/security/pam_howdy.so
+``` sh
+meson install -C build
 ```
 
-Change PAM config line to:
+Add the following line to your PAM configuration (/etc/pam.d/your-service):
 
-```pam
-auth     sufficient pam_howdy.so
+``` pam
+auth  sufficient  pam_howdy.so
 ```
