--- src/pam/.gitignore
@@ -1,1 +1,0 @@
-subprojects/*/
